made the class Clock public; Clock -> Clock.java
